requirements.txt
connexion[swagger-ui]==2.14.2
Flask==2.2.5
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
psycopg2==2.9.6
python-dotenv==1.0.0
SQLAlchemy==2.0.15